Clean up part 2.
