+++ cardano-node/app/chairman.hs
@@ -19,7 +19,8 @@ import           Ouroboros.Consensus.Protocol.Abstract (SecurityParam (..))
 import           Ouroboros.Consensus.Node.ProtocolInfo.Abstract (NumCoreNodes (..))
 import           Ouroboros.Consensus.NodeId (CoreNodeId)
 
-import           Cardano.Node.CLI
+import           Cardano.Common.CommonCLI (CommonCLI, mergeConfiguration, parseCommonCLI)
+import           Cardano.Common.Protocol (Protocol, SomeProtocol(..), fromProtocol)
 import           Cardano.Node.Parsers (parseCoreNodeId, parseProtocol)
 
 import           Cardano.Chairman (runChairman)
